作品情報ページのコンポーネント分割完了
