+++ src/components/atoms/index.tsx
@@ -1,7 +1,9 @@
 export { Button } from './Button';
 export { CategoryTitle } from './CategoryTitle';
+export { Img } from './Img';
 export { ItemList } from './ItemList';
 export { MovieTitle } from './MovieTitle';
 export { PointButton } from './PointButton';
 export { PrimaryText } from './PrimaryText';
 export { ScrollHint } from './ScrollHint';
+export { YouTube } from './YouTube';
