+++ src/components/organisms/index.tsx
@@ -2,3 +2,4 @@ export { AuthButton } from './AuthButton';
 export { Banner } from './Banner';
 export { Header } from './Header';
 export { MoviesList } from './MoviesList';
+export { TwoColumnInfo } from './TwoColumnInfo';
